+++ dimas/src/com/liveliness.rs
@@ -185,7 +185,8 @@ async fn run_liveliness<P>(
 				match sample.kind() {
 					SampleKind::Put => match p_cb.lock() {
 						Ok(mut lock) => {
-							if let Err(error) = lock(&ctx, id) {
+							let ctx = ctx.clone();
+							if let Err(error) = lock(ctx, id) {
 								error!("liveliness put callback failed with {error}");
 							}
 						}
@@ -197,7 +198,8 @@ async fn run_liveliness<P>(
 						if let Some(cb) = d_cb.clone() {
 							match cb.lock() {
 								Ok(mut lock) => {
-									if let Err(err) = lock(&ctx, id) {
+									let ctx = ctx.clone();
+									if let Err(err) = lock(ctx, id) {
 										error!("liveliness delete callback failed with {err}");
 									}
 								}
@@ -241,7 +243,8 @@ async fn run_initial<P>(
 						};
 						match p_cb.lock() {
 							Ok(mut lock) => {
-								if let Err(error) = lock(&ctx, id) {
+								let ctx = ctx.clone();
+								if let Err(error) = lock(ctx, id) {
 									error!("lveliness put callback failed with {error}");
 								}
 							}