+++ examples/liveliness/src/liveliness.rs
@@ -10,7 +10,7 @@ struct AgentProps {
 	num: u32,
 }
 
-fn liveliness_subscription(ctx: &Context<AgentProps>, id: &str) -> Result<()> {
+fn liveliness_subscription(ctx: Context<AgentProps>, id: &str) -> Result<()> {
 	println!("{id} is alive");
 	let mut val = ctx.read()?.num;
 	val += 1;
@@ -19,7 +19,7 @@ fn liveliness_subscription(ctx: &Context<AgentProps>, id: &str) -> Result<()> {
 	Ok(())
 }
 
-fn delete_subscription(ctx: &Context<AgentProps>, id: &str) -> Result<()> {
+fn delete_subscription(ctx: Context<AgentProps>, id: &str) -> Result<()> {
 	println!("{id} died");
 	let mut val = ctx.read()?.num;
 	val -= 1;